TS, alias, fixes

+++ src/helpers/RemoveFormCartButton/RemoveFormCartButton.jsx
@@ -4,8 +4,12 @@ import { useDispatch, useSelector } from 'react-redux';
 import { removeFromCart } from '../../redux/slices/cartSlice';
 import { useSendCartMutation } from '../../redux/api/cartEndpoints';
 import { getTokenFromCookies } from '../cookies/cookies';
+import { useModal } from '../../context/ModalContext';
 
-const RemoveFromCartButton = ({ product, children }) => {
+
+const RemoveFromCartButton = ({ product, withConfirmation = false, children }) => {
+
+  const { showModal } = useModal();
   const token = getTokenFromCookies();
   const dispatch = useDispatch();
 
@@ -13,12 +17,27 @@ const RemoveFromCartButton = ({ product, children }) => {
 
   const handleRemoveFromCartClick = (e) => {
     e.preventDefault();
-    // dispatch(removeFromCart(product));
+
+    if (withConfirmation) {
+      showModal({type: 'confirmation', 'title': 'Удаление', 'text': 'Удалить товар из корзины?', action: remove, product: product});
+      return
+    } 
+
+
     if (token) {
       sendCart({ id: product.id, quantity: 0, selected: 0 })
     } 
     dispatch(removeFromCart(product));
   };
+
+  const remove = (product) => {
+    
+    if (token) {
+      sendCart({ id: product.id, quantity: 0, selected: 0 })
+    } 
+    dispatch(removeFromCart(product));
+  }
+
 
   return children({ isLoading, handleRemoveFromCartClick });
 };
